version 1.1
feat: 添加多语种学习平台后台管理系统

refactor: 重构用户服务实现以匹配框架安全模块

fix: 修正语音识别评分计算精度问题

style: 优化停止脚本的输出格式和编码处理

docs: 添加教育管理菜单初始化SQL和后台管理指南

chore: 更新前端代理配置指向本地开发环境

build: 添加新的HTML模板和Vue组件文件

test: 移除冗余的Web配置类

perf: 优化学习路径推荐算法的数据类型处理

ci: 更新启动脚本以支持多服务并行启动

+++ ruoyi-system/src/main/java/com/ruoyi/edu/service/impl/SpeechRecognitionServiceImpl.java
@@ -483,7 +483,7 @@ public class SpeechRecognitionServiceImpl implements ISpeechRecognitionService {
                 }
             }
 
-            wordScores.put(refWord, Math.round(score));
+            wordScores.put(refWord, (double) Math.round(score));
         }
 
         return wordScores;
